delivery almost done

+++ src/main/java/kea/grocery/reposities/DeliveryRepository.java
@@ -3,5 +3,6 @@ package kea.grocery.reposities;
 import kea.grocery.entities.Delivery;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DeliveryRepository extends JpaRepository<Delivery, Integer> {
+public interface DeliveryRepository extends JpaRepository<Delivery, Long> {
+
 }
