+++ src/main/java/kea/grocery/reposities/ProductOrderRepository.java
@@ -3,5 +3,5 @@ package kea.grocery.reposities;
 import kea.grocery.entities.ProductOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductOrderRepository extends JpaRepository<ProductOrder, Integer> {
+public interface ProductOrderRepository extends JpaRepository<ProductOrder, Long> {
 }
