+++ src/main/java/kea/grocery/reposities/VanRepository.java
@@ -3,5 +3,5 @@ package kea.grocery.reposities;
 import kea.grocery.entities.Van;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VanRepository extends JpaRepository<Van, Integer> {
+public interface VanRepository extends JpaRepository<Van, Long> {
 }
